Replace hardcoded vote ajax with partials
Instead of writing helper, coffeescript, views etc. for each votable
we now have one partial to rule them all. When clicked the partial
posts to an endpoint depending on the votable name, processing the vote
and returning dynamic js to update/rerender itself. This allows for
ajax voting on any new model by simply adding a route and a controller
method

Signed-off-by: Tristan <tristan.schoenhals@gmail.com>